Add files via upload
A Single Product Billing System is a C-based console application that calculates the total cost of one product using its price and quantity. It generates a simple bill and it check the value is positive or negative and compered the product it's all of  understand basic C concepts like conditions, loops, and arithmetic operations.